created volume entity and revised product controller
In addition to creating a Volume entity class that shares a relationship with Product entities, I also revised how the product controller returns data. I have added javascript to the client side html file so that rendered html fragments are requested from the sever in response to user interactions. Now filtering of the product table and creating/updating a product no longer intervene with the other action. Next step is to redo controller testing accordingly and add a functionality for create, update, and delete of the Volume entity