update to attributes page and custom login page
I created the attributes page to include both the categories and volumes. I updated user interactions on this page to function similar to the products page. I also added a custom login page and a logout button on the nav-bar. The next step is to update the controller testing for all current controllers