implemented soft deletes
I have switched to soft deletes for product, category, and volume entities. This ensures that data integrity is maintained in the database. Entities with this boolean value set to true are not visible by default, and I plan to only allow admins to view or these"deleted" items and have the ability to "reactivate" them. I also implemented pagination for categories and volumes.