added MySQL connectivity along with custom users and user roles
I created an application profile to enable a datasource connection to a local MySQL instance. Then, I used flyway to initialize the database schema for this application. I then created a custom user entity along with a custom user details service, and configured spring security to use this implementation. UserRoles were then refactored from strings to an enums to work with the new entity. The next step is to create the next flyway migration script to create the "users" table in the local MySQL instance