Clean up exceptions and add Javadoc for services
I have added Javadoc for the service interfaces, and also for implementing methods with particular functionality not described in the interfaces. I have also removed the ItemDeleteException class and made the ItemCreationException a checked exception. I have left the NotFoundException class unchecked for the controller to decied wheather to handle it or not